Change setSecurityStatus method to copySecurityIfVerified
