--- apps/server/src/modules/files-storage/helper/scan-status.ts
@@ -1,7 +1,0 @@
-import { FileRecord, FileSecurityCheck } from '../entity';
-
-export function deriveStatusFromSource(sourceFile: FileRecord, targetFile: FileRecord): FileSecurityCheck {
-	const securityCheck = sourceFile.isVerified() ? sourceFile.securityCheck : targetFile.securityCheck;
-
-	return securityCheck;
-}
